Refactor: profile.dart 이름 user_info.dart 로 변경
